Add the standard operator package and the serialisation codes into the python core
This patch temporarily includes all the standard operators and serialization codecs in the Python API core dependencies because the initialization subsystem does not support dynamic loading of new operators/codecs from Python extensions.

This logic will require revision in the future. Particularlly, it should be possible to launch Arolla with a minimal set of operators to define a custom operator library.

PiperOrigin-RevId: 621859947
Change-Id: Ief7db4ff90f74c163954e0ebecbd124e64990642