Enable bazel test //... for macos
This change adds a small number of AROLLA_API annotations to make all tests pass on macOS. (Further API cleanup will be needed in the future.)

We should also consider introducing `arolla_cc_library` to better control the `AROLLA_SHARED_COMPILATION` state. Currently, the shared library mode, which is only required for the Python API, is enabled for everything in `.bazelrc`.

PiperOrigin-RevId: 626433815
Change-Id: I657d2d9bd56ac156c9c1e3946cf76278f7c82b3b