Merge `//arolla/serialization:{decode,encode}`
Motivation: The current policy for the project requires a single "public" target per directory, and `//arolla/serialization` already has decode and encode. Now, we need to add utilities for "ExprSet" that don’t specifically belong to either of the existing targets. To avoid further stretching the policy, it would be better to merge the targets. Besides, we already have a single `//arolla/serialization_base`.

The plan is to completely merge "//arolla/serialization:decode" and "//arolla/serialization:encode" into "//arolla/serialization" in a follow change.

PiperOrigin-RevId: 631702271
Change-Id: I2f46f50e4fbb18620b202507257a93a5864a2403