Add support for sequence-enabled __getitem__ on ExprView, useful for unpacking tuple-like values.
To allow and ExprView to behave like a sequence instead of a mapping, implement `_arolla_sequence_getitem_`, which will be used as `__getitem__` when the Python interpreter treats the object as a sequence (e.g. by `iter()`). The separate method name makes it possible to implement `__getitem__` without opting into sequence-like behavior.

PiperOrigin-RevId: 652984796
Change-Id: Id924849ffe58035ca3228fd32b1009ae7a484ce5