Split `math.sum` operator into `math.sum` and `math._sum_spase`.
I remove `initial=` argument, because only two meaningful values were `0` and `None`. It allowed to simplify several use cases. For now I keep `math._sum_spase` private, as the only clients are RLv1 and a couple of other operators. Don't want to get accidental usages if we decide to move it to `legacy.*` namespace.

PiperOrigin-RevId: 653155716
Change-Id: I47a4c9f5a89a72f2e3ddd6ba0a049b91d55b7735