Provide a helper PyErr_CanCallCheckSignal()
Here, we introduce a helper function that allows to check whether `PyErr_CheckSignal()` can be used to check the signals within the current context. The motivation for this is to provide support for keyboard interruption during evaluation.

Unfortunately, we found no public API to implement such a check, so for now, we rely on the internal Python API.

PiperOrigin-RevId: 659913030
Change-Id: Id1533e2a71fa2b859bfe0581fa3ffd3737ada6d8