Rewrite detect_qtype_signatures in C++
The motivation for this change is to speed up the testing infrastructure for the operators.

Note: We still have the same exponential algorithm as before, which will not work for operators with large arity. We are keeping it because the rewriting in C++ can be done almost mechanically, and it already provides a visible speed-up.

We'll eventually need to replace it with a quasi-linear 'pinning' algorithm, which will require slightly more work to implement.

Benchmark results:
  * `%timeit len(arolla.testing.detect_qtype_signatures(arolla.M.core.presence_or))`
    * 7.86ms -> 1.35ms
  * `%timeit len(arolla.testing.detect_qtype_signatures(arolla.M.math.sigmoid))`
    * 537ms -> 212ms
  * `%timeit len(arolla.testing.detect_qtype_signatures(arolla.M.bool.logical_if))`
    * 29.5s -> 12.1s

PiperOrigin-RevId: 661195816
Change-Id: I230ebc309caa04247f638b8edbcf89e181e0e858