Handle inputs with NOTHING type before testing the qtype constraint
Motivation: We want to reserve the NOTHING type to indicate if the type of a parameter is not yet known. Thus, we want to differentiate it from cases where the type of the parameter is known to be NOTHING.

This change should be safe because NOTHING has no values and can never be an actual input during the computation.

PiperOrigin-RevId: 663738415
Change-Id: I9daa1db71540a36a142758c5dbef9d1a307689a3