Stop using QExpr operator names outside of QExpr directory
It seems that there are no important usages of QExpr operator names left. If we keep the names only in the registry, it will allow to simplify operator definitions, i.e. one won't need to provide the names in two separate places. See cl/669008221 with the simplification.

The cost of the change is slightly less informative error messages. But the messages in question are very low level, and we add the details about the operators up the stack anyway.

The change is motivated by the Felix's work on variadic operator family, it will e.g. simplify cl/668902341 a bit.

PiperOrigin-RevId: 669377213
Change-Id: Ic110479c5fbe4c149e722b0bbdbbfe32feba3d41