Support storing additional information in `signature.aux_policy`
This patch enables storing additional information in `signature.aux_policy` after the ":" (e.g., `aux_policy = "aux_policy_name:aux_policy_param"`). This additional information will not be used in aux-policy lookup (it is exclusively for the policy's internal use) and will not affect the serialisation format (as it is stored as part of the "signature_spec").

Motivation: The goal is to enable more advanced behaviour for Arolla operators in Python, such as supporting positional-only, keyword-only, and variadic-keyword parameters, as well as parameter-specific "boxing" logic.
PiperOrigin-RevId: 698371641
Change-Id: I2371d9c88436f4878359897d4c66bfc132824e9f