Migrate to InferAttrs instead of GetOutputQType in StdFunctionOperator
Why? It allows us to fail early if some inputs are known to be wrong while others are still unknown. This greatly speeds up detect_qtype_signatures, as seen in [1].

NOTE: I've chosen to _not_ change the user-provided inference fn from a qtype inference fn to an attr inference fn because it then becomes much easier to accidentally write an attr inference function that returns an ExprAttribute containing a Literal, causing the evaluation function to be skipped entirely. I prefer to not make this possible at all.
PiperOrigin-RevId: 702340211
Change-Id: I499b9669b87466e62d0787c6fd50d44d7d69d9f3