Add helpers arolla.optoosl.fix_trace_args_kwargs(...)
This patch adds helpers for the normalisation of `*args` and `**kwargs` during function tracing.

Motivation: The goal is to simplify the support for `**kwargs` in custom binding policies; and to unify the handling of variadic parameters in traceable code.

Additionally, it provides some flexibility to adjust the internal representation of `*args` and `**kwargs` in the future.

PiperOrigin-RevId: 704233723
Change-Id: Id2895af6eecfe651ce4b2d8b4b497372b5be42de