Restore passing `decayed_op` explicitly to the compiler extensions.
I replaced it with `node` in cl/706626794. All the operators that used extensions emerged in compile time, so it was not a big deal (except for a decision forest, but it is strange to register one as well). However for `StdFunctionOperator` it makes sense (cl/705853751). For consistency I migrate all the operators to use `decayed_op`.

PiperOrigin-RevId: 707048469
Change-Id: I68b8daeaea34949aa7749788b9f59b7d45f28b96