Prototype support for cancellation of aggregation operators
As a proof of concept, added cancellation support to the array.dense_rank operator.

This implementation checks for cancellation only in the `FinalizeFullGroup` method, skipping the `std::sort` call (the most expensive part of the operator). An alternative approach would be to check for cancellation within the template code that uses the operator, potentially allowing for even earlier termination.

PiperOrigin-RevId: 721323307
Change-Id: Ie8a72fc47617ccd200f532c3c2ef61c7438b6120