arolla::python::py_cancellation_controller
This CL provides a basic implementation of "py_cancellation_controller" to support `PyCancellationScope`. This implementation relies on installing a signal handler during initialisation.

The implementation is easily extensible, particularly to provide a mechanism
for re-installing a signal handler at runtime and/or switching to using
`PySignal_SetWakeupFd` or `signalfd`. However, we are starting with this
base version as it is the most reliable, and we expect it to be sufficient
for most applications.

PiperOrigin-RevId: 739198177
Change-Id: Ic866f20a2c25d58951816a05bdd46f9916a59efb