Address compatibility with GCC 12
Apparently, `AROLLA_API` and `[[nodiscard]]` cannot be specified simultaneously in g++-12, causing an `error: expected identifier before '__attribute__'`, that looks like a syntactic error. As a workaround, we specify `[[nodiscard]]` on the forward declaration (which is sufficient) and `AROLLA_API` on the definition.

This problem does not occur in later compiler versions, but we need GCC 12 for compatibility with Google Colab.

PiperOrigin-RevId: 745654199
Change-Id: I96d50f220cdfa923b5d6f1bbbd2609266edb6549