Improve the error message for when inputs are accidentally captured into py-fns
I've included the expr that fails to help with debugging. Additionally, for most kd.py.[...] operators, I also include the user-provided py function since this is not shown otherwise. For kd.py.map_py_on_cond, I have not done so, since there the responsible function depends on which variable is the failing one. This is a bit hard to figure out, so I've skipped this for now.

PiperOrigin-RevId: 748658374
Change-Id: I565d15f9a4397fc37b5969bbb5ff88a1a053b732