Refactor PyTuple_AsSpan
Motivation: While `PySequence_Fast` is documented to return the object itself if it is already a tuple or list, it is ambiguous whether this applies only to exact type matches or also includes subtypes (and the implementation actually performs exact type checks).

The new implementation of `PyTuple_AsSpan` depends solely on the fact that tuples and lists store their items in a contiguous internal array (that `PySequence_Fast_ITEMS` also depends on) but uses `Py{Tuple,List}_GET_ITEM`, which allows subtypes.

PiperOrigin-RevId: 750119779
Change-Id: Ifa55712c0b4a17fb224ff80819d3542b13cbcc04