Labelled QType
Labelled QType is a lightweight implementation of the Derived QType
interface that can be dynamically instantiated, allowing creation of new
types without re-compilation.

Additionally, we are moving the M.derived_qtype.* namespace into
a separate package. The main motivation here is to limit the complexity
of the core library — new features based on the derived types should
likely be implemented outside of the core.

PiperOrigin-RevId: 772647285
Change-Id: I48ea3889394ffad3cdda59c709ca4fee20ef2cf6