Improve the lazy initialisation for attributes
Previously, the following pattern was used for lazy initialisation:

```python
  @property
  def attr(self)
    try:
      return self._attr
    except AttributeError:
      self._attr = init_attr(...)
      return self._attr
```

However, if `init_attr(...)` failed, the error would appear to be related to/caused by an `AttributeError`, which was misleading.

This patch switches to the following pattern:
```python
  @property
  def attr(self)
    try:
      return self._attr
    except AttributeError:
      pass
    self._attr = init_attr(...)
    return self._attr
```
to make the intialisation error less confusing.

(Note: Our experiments has shown that `AttributeError` is still a faster option than a `hasattr` check for the lazy initalisation.)

PiperOrigin-RevId: 793627775
Change-Id: Ia6456520c8201def9e45c3b9990a44b71469e2bf